Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main() {
-    float a;
-    scanf("%0.4f",&a);
-    printf("You entered: %0.4f",a);
+    double a;
+    scanf("%0.4lf",&a);
+    printf("You entered: %0.4lf",a);
     return 0;
 }
